float -> double for better precision

+++ main.cpp
@@ -9,9 +9,9 @@ int main() {
     clock_t start_2, finish_2;
     clock_t start_3, finish_3;
 
-    long long fib1;
-    long long fib2;
-    long long fib3;
+    long long int fib1;
+    long long int fib2;
+    long long int fib3;
 
     int choice;
     do {
@@ -26,7 +26,7 @@ int main() {
         fib1 = fib_recursive(choice);
         cout << iteration_recursive << "\t\tRECURSIVE\t";
         finish_1 = clock() - start_1;
-        cout << (float) finish_1 / CLOCKS_PER_SEC << "s.\t\t" << fib1 << endl;
+        cout << (double) finish_1 / CLOCKS_PER_SEC << "s.\t\t" << fib1 << endl;
 
         //------------------ Iterative Fibonacci ------------------
 
@@ -34,15 +34,15 @@ int main() {
         fib2 = fib_iterative(choice);
         cout << iteration_iterative << "\t\tITERATIVE\t";
         finish_2 = clock() - start_2;
-        cout << (float) finish_2 / CLOCKS_PER_SEC << "s.\t\t" << fib2 << endl;
+        cout << (double) finish_2 / CLOCKS_PER_SEC << "s.\t\t" << fib2 << endl;
 
-        //------------------ Binary Fibonacci ------------------
+        //------------------ Matrix exponentiation Fibonacci ------------------
 
         start_3 = clock();
         fib3 = fib_binary(choice);
         cout << iteration_binary << "\t\tBINARY\t\t";
         finish_3 = clock() - start_3;
-        cout << (float) finish_3 / CLOCKS_PER_SEC << "s.\t\t" << fib3 << endl;
+        cout << (double) finish_3 / CLOCKS_PER_SEC << "s.\t\t" << fib3 << endl;
 
         cout << "---------------------------------------------------------" << endl;
